fuzzystrmatch levenshtein function:
Use a levenshtein algorithm with the possibility of using a threshold.
Algorithm will only explore a "stripe" of possible edit distances within
a given threshold and give up (possibly early) if it can't find a match
within this threshold.

Written against 8.4 before I knew (or maybe before) similar work was
going on in mainline.